CREATE TABLE `logistdb`.`price_protocol` (
  `idprice_protocol` INT NOT NULL AUTO_INCREMENT,
  `barcode` VARCHAR(255) COMMENT 'Штрих-код товара',
  `product_code` VARCHAR(100) COMMENT 'Код товара',
  `tnv_code` VARCHAR(100) COMMENT 'Код ТНВЭД (товарной номенклатуры внешнеэкономической деятельности)',
  `name` VARCHAR(500) COMMENT 'Наименование товара',
  `price_producer` DECIMAL(10,2) COMMENT 'Прейскурантная цена производителя без НДС (BYN)',
  `cost_importer` DECIMAL(10,2) COMMENT 'Себестоимость импортера без надбавки и без НДС (гр.11 ТТН)',
  `markup_importer_percent` DECIMAL(5,2) COMMENT 'Надбавка импортера, % (гр.11 ТТН)',
  `discount_percent` DECIMAL(5,2) COMMENT 'Скидка с отпускной цены, %',
  `wholesale_discount_percent` DECIMAL(5,2) COMMENT 'Оптовая скидка, %',
  `price_without_vat` DECIMAL(10,2) COMMENT 'Отпускная цена без НДС (гр. 4 ТТН), BYN',
  `wholesale_markup_percent` DECIMAL(5,2) COMMENT 'Оптовая надбавка, %',
  `vat_rate` DECIMAL(5,2) COMMENT 'Ставка НДС',
  `price_with_vat` DECIMAL(10,2) COMMENT 'Цена с НДС, BYN',
  `country_origin` VARCHAR(255) COMMENT 'Страна происхождения',
  `manufacturer` VARCHAR(255) COMMENT 'Производитель',
  `unit_per_pack` VARCHAR(100) COMMENT 'Штук/кг в упаковке',
  `shelf_life_days` INT COMMENT 'Срок годности в днях',
  `current_price` DECIMAL(10,2) COMMENT 'Текущая отпускная цена, BYN',
  `price_change_percent` DECIMAL(5,2) COMMENT '% изменения отпускной цены',
  `last_price_change_date` DATE COMMENT 'Дата последнего изменения цены',
  `valid_from` DATE COMMENT 'Дата начала действия',
  `valid_to` DATE COMMENT 'Дата окончания действия',
  `contract_number` VARCHAR(100) COMMENT 'Номер договора',
  `contract_date` DATE COMMENT 'Дата договора',
  PRIMARY KEY (`idprice_protocol`))
COMMENT = 'протокол согласования цены';


ALTER TABLE `logistdb`.`telegram_chat_quality` 
CHANGE COLUMN `chat_id` `chat_id` BIGINT NOT NULL ;

+++ src/main/java/by/base/main/dao/impl/TelegramChatQualityDAOImpl.java
@@ -36,7 +36,7 @@ public class TelegramChatQualityDAOImpl implements TelegramChatQualityDAO{
 	
 	private static final String queryGetListObjByUser = "from TelegramChatQuality where chatId=:chatId";
 	@Override
-	public boolean existsById(int chatId) {
+	public boolean existsById(Long chatId) {
 		Session currentSession = sessionFactory.getCurrentSession();		
 		Query<TelegramChatQuality> theObject = currentSession.createQuery(queryGetListObjByUser, TelegramChatQuality.class);
 		theObject.setParameter("chatId", chatId);		
@@ -55,7 +55,7 @@ public class TelegramChatQualityDAOImpl implements TelegramChatQualityDAO{
 	
 	private static final String queryDeleteById = "delete from TelegramChatQuality where chatId=:chatId";
 	@Override
-	public void deleteByChatId(int chatId) {
+	public void deleteByChatId(Long chatId) {
 		Session currentSession = sessionFactory.getCurrentSession();
 		Query theQuery = 
 				currentSession.createQuery(queryDeleteById);
